requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
greenlet==3.1.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
m3u-ipytv==0.2.11
m3u8==6.0.0
mpv==1.0.7
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-mpv==1.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
